unify and simplify user space makefiles rename NETWORKING configuration option to CONFIG_NETWORKING

+++ uspace/lib/libc/arch/ia64/include/types.h
@@ -37,7 +37,7 @@
 
 #define __64_BITS__
 
-typedef unsigned long long sysarg_t;
+typedef unsigned long sysarg_t;
 
 typedef char int8_t;
 typedef short int int16_t;
